Backend finished

+++ back-end/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 //@desc Get goals
 //@route GET
 //@access Private
 const getGoals = asyncHandler(async(req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id })
     if(goals == null){
         res.status(400).json({message:"No goals are found!"});
     }
@@ -21,7 +22,8 @@ const setGoals = asyncHandler(async(req,res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
     res.status(200).json(goal);
 })
@@ -34,6 +36,20 @@ const updateGoals = asyncHandler(async(req,res) => {
     if(!goal){
         res.status(400) 
         throw new Error('Goal not found!');
+    }
+
+    //Check user exist
+    const user = await User.findById(req.user.id);
+
+    if(!user){
+        res.status(401)
+        throw new Error("User not found!");
+    }
+
+    //Make sure login user is macthed
+    if(goal.user.toString()!== user.id){
+        res.status(400)
+        throw new Error("User not matched!");
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,{
         new:true,
@@ -49,6 +65,20 @@ const deleteGoals = asyncHandler(async(req,res) => {
     if(!goal){
         res.status(400).json({message:"Goal not found!"});
     }
+
+     //Check user exist
+     const user = await User.findById(req.user.id);
+
+     if(!user){
+         res.status(401)
+         throw new Error("User not found!");
+     }
+ 
+     //Make sure login user is macthed
+     if(goal.user.toString()!== user.id){
+         res.status(400)
+         throw new Error("User not matched!");
+     }
 
     await goal.deleteOne();
     res.status(200).json({id:req.params.id});
